# frontend/.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NeuLitTrace API",
    "version": "0.1.0"
  },
  "paths": {
    "/query": {
      "post": {
        "summary": "Query",
        "operationId": "query_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "400": {
            "description": "the request body could not be decoded",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string",
                      "title": "Detail"
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "title": "ParseErrorResponse",
                  "description": "Starlette's own shape for an undecodable body, written down."
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/query/stream": {
      "post": {
        "summary": "Query Stream",
        "description": "Runs the same pipeline as POST /query but over SSE, emitting `stage`\nevents as the search loop progresses (hyde_expand, retrieval,\nrelevance_check, refine_query, summarize, citation_check - the same\nstage names v1 used) followed by one `done` event with the full result.\nThe frozen HTTP contract in plan-v2/00-SHARED-CONTRACTS.md section 4\ndoesn't define this route's event shape, so it's ours to choose; matching\nv1's stage names is what frontend/src/components/progress-timeline.tsx\nalready expects.",
        "operationId": "query_stream_query_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "server-sent `stage` events followed by one `done` event",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                },
                "itemSchema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "string",
                      "contentMediaType": "application/json",
                      "contentSchema": {
                        "anyOf": [
                          {
                            "type": "object",
                            "properties": {
                              "type": {
                                "const": "stage"
                              },
                              "stage": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "type",
                              "stage"
                            ]
                          },
                          {
                            "type": "object",
                            "properties": {
                              "type": {
                                "const": "done"
                              },
                              "result": {
                                "$ref": "#/components/schemas/QueryResponse"
                              }
                            },
                            "required": [
                              "type",
                              "result"
                            ]
                          },
                          {
                            "type": "object",
                            "properties": {
                              "type": {
                                "const": "error"
                              },
                              "message": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "type",
                              "message"
                            ]
                          }
                        ]
                      }
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "the request body could not be decoded",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string",
                      "title": "Detail"
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "title": "ParseErrorResponse",
                  "description": "Starlette's own shape for an undecodable body, written down."
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/memory/profile": {
      "get": {
        "summary": "Memory Profile",
        "operationId": "memory_profile_memory_profile_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/memory/specialty": {
      "post": {
        "summary": "Memory Specialty",
        "operationId": "memory_specialty_memory_specialty_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpecialtyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "the request body could not be decoded",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string",
                      "title": "Detail"
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "title": "ParseErrorResponse",
                  "description": "Starlette's own shape for an undecodable body, written down."
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/memory/forget": {
      "post": {
        "summary": "Memory Forget",
        "operationId": "memory_forget_memory_forget_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "the request body could not be decoded",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string",
                      "title": "Detail"
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "title": "ParseErrorResponse",
                  "description": "Starlette's own shape for an undecodable body, written down."
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/memory/thread": {
      "get": {
        "summary": "Memory Thread",
        "operationId": "memory_thread_memory_thread_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BrainRegionOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "atlas_label": {
            "type": "string",
            "title": "Atlas Label"
          },
          "region_literature": {
            "type": "string",
            "title": "Region Literature"
          }
        },
        "type": "object",
        "required": [
          "name",
          "atlas_label",
          "region_literature"
        ],
        "title": "BrainRegionOut"
      },
...
      "CallSiteCostOut": {
        "properties": {
          "tokens": {
            "type": "integer",
            "title": "Tokens"
          },
          "cost_usd": {
            "type": "number",
            "title": "Cost Usd"
          }
        },
        "type": "object",
        "required": [
          "tokens",
          "cost_usd"
        ],
        "title": "CallSiteCostOut"
      },
      "CitationOut": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "pmid": {
            "type": "string",
            "title": "Pmid"
          },
          "supported": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supported"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": ""
          },
          "retrieval_rank": {
            "type": "integer",
            "title": "Retrieval Rank",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "index",
          "pmid",
          "supported",
          "note"
        ],
        "title": "CitationOut",
        "description": "`index` is the [N] slot in the prompt the summary was written against.\n`retrieval_rank` and `source` are the retrieval provenance: the position\nthe record held in the ordering the answer was built from, and the backend\nthat served it. Both are carried through rather than re-derived, so a\nstored answer can still name its sources after the prompt is gone."
      },
...
      "CostOut": {
        "properties": {
          "total_tokens": {
            "type": "integer",
            "title": "Total Tokens"
          },
          "cost_usd": {
            "type": "number",
            "title": "Cost Usd"
          },
          "by_call_site": {
            "additionalProperties": {
              "$ref": "#/components/schemas/CallSiteCostOut"
            },
            "type": "object",
            "title": "By Call Site"
          }
        },
        "type": "object",
        "required": [
          "total_tokens",
          "cost_usd",
          "by_call_site"
        ],
        "title": "CostOut"
      },
...
      "ForgetRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "ForgetRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemoryOut": {
        "properties": {
          "applied": {
            "type": "boolean",
            "title": "Applied"
          },
          "seen_filtered": {
            "type": "integer",
            "title": "Seen Filtered"
          },
          "profile_used": {
            "type": "boolean",
            "title": "Profile Used"
          },
          "distilled_context": {
            "type": "string",
            "title": "Distilled Context"
          }
        },
        "type": "object",
        "required": [
          "applied",
          "seen_filtered",
          "profile_used",
          "distilled_context"
        ],
        "title": "MemoryOut"
      },
...
      "ProfileOut": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "specialty": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialty"
          },
          "conditions_explored": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Conditions Explored"
          },
          "query_count": {
            "type": "integer",
            "title": "Query Count"
          },
          "distilled_context": {
            "type": "string",
            "title": "Distilled Context"
          },
          "seen_pmid_count": {
            "type": "integer",
            "title": "Seen Pmid Count"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "specialty",
          "conditions_explored",
          "query_count",
          "distilled_context",
          "seen_pmid_count"
        ],
        "title": "ProfileOut"
      },
      "QueryRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Query"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "personalize": {
            "type": "boolean",
            "title": "Personalize",
            "default": false
          },
          "policy": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "tight",
                  "generous"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy"
          }
        },
        "type": "object",
        "required": [
          "query",
          "session_id",
          "user_id"
        ],
        "title": "QueryRequest"
      },
      "QueryResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "summary_markdown": {
            "type": "string",
            "title": "Summary Markdown"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationOut"
            },
            "type": "array",
            "title": "Citations"
          },
          "papers": {
            "items": {
              "$ref": "#/components/schemas/ScoredPaperOut"
            },
            "type": "array",
            "title": "Papers"
          },
          "trace": {
            "items": {
              "$ref": "#/components/schemas/TraceRoundOut"
            },
            "type": "array",
            "title": "Trace"
          },
          "region": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BrainRegionOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "memory": {
            "$ref": "#/components/schemas/MemoryOut"
          },
          "cost": {
            "$ref": "#/components/schemas/CostOut"
          },
          "policy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PolicyOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "grounding": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GroundingOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "retrieval_provenance": {
            "items": {
              "$ref": "#/components/schemas/RecordProvenanceOut"
            },
            "type": "array",
            "title": "Retrieval Provenance",
            "default": []
          },
          "abstained": {
            "type": "boolean",
            "title": "Abstained",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "summary_markdown",
          "citations",
          "papers",
          "trace",
          "region",
          "memory",
          "cost"
        ],
        "title": "QueryResponse"
      },
...
      "ScoredPaperOut": {
        "properties": {
          "paper": {
            "$ref": "#/components/schemas/PaperOut"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "lexicalScore": {
            "type": "number",
            "title": "Lexicalscore"
          },
          "semanticScore": {
            "type": "number",
            "title": "Semanticscore"
          },
          "rarityMultiplier": {
            "type": "number",
            "title": "Raritymultiplier"
          },
          "memoryMultiplier": {
            "type": "number",
            "title": "Memorymultiplier"
          }
        },
        "type": "object",
        "required": [
          "paper",
          "score",
          "lexicalScore",
          "semanticScore",
          "rarityMultiplier",
          "memoryMultiplier"
        ],
        "title": "ScoredPaperOut"
      },
...
      "SpecialtyRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "specialty": {
            "type": "string",
            "title": "Specialty"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "specialty"
        ],
        "title": "SpecialtyRequest"
      },
      "ThreadOut": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "queries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Queries"
          },
          "pmids_shown": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pmids Shown"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "user_id",
          "queries",
          "pmids_shown"
        ],
        "title": "ThreadOut"
      },
      "TraceRoundOut": {
        "properties": {
          "iteration": {
            "type": "integer",
            "title": "Iteration"
          },
          "retrieved_pmids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Retrieved Pmids"
          },
          "relevant": {
            "type": "boolean",
            "title": "Relevant"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "note": {
            "type": "string",
            "title": "Note"
          },
          "memory_applied": {
            "type": "boolean",
            "title": "Memory Applied"
          },
          "seen_filtered": {
            "type": "integer",
            "title": "Seen Filtered"
          }
        },
        "type": "object",
        "required": [
          "iteration",
          "retrieved_pmids",
          "relevant",
          "confidence",
          "note",
          "memory_applied",
          "seen_filtered"
        ],
        "title": "TraceRoundOut"
      },
...
    }
  }
}